@@ -1,29 +1,26 @@
 {
-  "_from": "raw-body@2.3.2",
-  "_id": "raw-body@2.3.2",
+  "_from": "raw-body@2.3.3",
+  "_id": "raw-body@2.3.3",
   "_inBundle": false,
-  "_integrity": "sha1-vNYMd9Prk83gBQKVw/N5OJvIj4k=",
+  "_integrity": "sha512-9esiElv1BrZoI3rCDuOuKCBRbuApGGaDPQfjSflGxdy4oyzqghxu6klEkkVIvBje+FF0BX9coEv8KqW6X/7njw==",
   "_location": "/raw-body",
-  "_phantomChildren": {
-    "inherits": "2.0.3",
-    "statuses": "1.4.0"
-  },
+  "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "raw-body@2.3.2",
+    "raw": "raw-body@2.3.3",
     "name": "raw-body",
     "escapedName": "raw-body",
-    "rawSpec": "2.3.2",
+    "rawSpec": "2.3.3",
     "saveSpec": null,
-    "fetchSpec": "2.3.2"
+    "fetchSpec": "2.3.3"
   },
   "_requiredBy": [
     "/body-parser"
   ],
-  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.3.2.tgz",
-  "_shasum": "bcd60c77d3eb93cde0050295c3f379389bc88f89",
-  "_spec": "raw-body@2.3.2",
+  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.3.3.tgz",
+  "_shasum": "1b324ece6b5706e153855bc1148c65bb7f6ea0c3",
+  "_spec": "raw-body@2.3.3",
   "_where": "C:\\Users\\Auxiliar admon\\Desktop\\sento\\node_modules\\body-parser",
   "author": {
     "name": "Jonathan Ong",
@@ -46,25 +43,25 @@
   ],
   "dependencies": {
     "bytes": "3.0.0",
-    "http-errors": "1.6.2",
-    "iconv-lite": "0.4.19",
+    "http-errors": "1.6.3",
+    "iconv-lite": "0.4.23",
     "unpipe": "1.0.0"
   },
   "deprecated": false,
   "description": "Get and validate the raw body of a readable stream.",
   "devDependencies": {
-    "bluebird": "3.5.0",
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "bluebird": "3.5.1",
+    "eslint": "4.19.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.11.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.7.0",
+    "eslint-plugin-standard": "3.1.0",
     "istanbul": "0.4.5",
     "mocha": "2.5.3",
-    "readable-stream": "2.3.3",
-    "safe-buffer": "5.1.1"
+    "readable-stream": "2.3.6",
+    "safe-buffer": "5.1.2"
   },
   "engines": {
     "node": ">= 0.8"
@@ -89,5 +86,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --trace-deprecation --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --trace-deprecation --reporter spec --check-leaks test/"
   },
-  "version": "2.3.2"
+  "version": "2.3.3"
 }
